pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ezcomputerctrl"
version = "0.1.0"
description = "面向 GUI 语义执行的第一阶段最小闭环实现"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.19.0",
    "openai>=1.75.0",
    "pillow>=11.2.0",
    "pydantic>=2.12.0",
    "watchdog>=6.0.0",
]

[project.scripts]
ezcomputerctrl-mcp = "ezcomputerctrl.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
